Updated the aspirin driver to the newest version. Added MPU6000 reset at init.
Also when we are not getting any data from the MPU6000 for some time we
will now try to reinitialize it.

The delay in main.c is disabled for now as it does not seem to be needed
any more. (cross fingers) Please delete it for good after validating
this patch! :)